[정주원] #2 - 메인 화면 DB에서 값 받아와서 구현 중 #7 - 리뷰 삭제 기능 추가 -

+++ src/main/java/com/nowjoo/nowjiu/administrator/dto/AdministratorMainDto.java
@@ -7,8 +7,8 @@ import lombok.Getter;
 @Getter
 public class AdministratorMainDto {
 
-	private int userCount;
-	private int goodsCount;
-	private int orderCount;
-	private int reviewCount;
+	private long userCount;
+	private long goodsCount;
+	private long orderCount;
+	private long reviewCount;
 }
